Fixed Signed integer overflow

+++ src/doom/p_enemy.cpp
@@ -860,7 +860,7 @@ void A_CPosAttack(mobj_t *actor)
     bangle = actor->angle;
     slope  = P_AimLineAttack(actor, bangle, MISSILERANGE);
 
-    angle  = bangle + (P_SubRandom() << 20);
+    angle  = static_cast<int>(static_cast<int64_t>(bangle) + static_cast<int64_t>(P_SubRandom() << 20));
     damage = ((P_Random() % 5) + 1) * 3;
     P_LineAttack(actor, angle, MISSILERANGE, slope, damage);
 }
